+++ flytekit/control_plane/tasks/task.py
@@ -1,16 +1,35 @@
+from typing import Any, Optional, Union
+
 from flytekit.common.exceptions import scopes as _exception_scopes
+from flytekit.common.exceptions import user as _user_exceptions
 from flytekit.common.mixins import hash as _hash_mixin
 from flytekit.control_plane import identifier as _identifier
 from flytekit.control_plane import interface as _interfaces
+from flytekit.core.base_task import ExecutableTaskMixin
+from flytekit.core.context_manager import FlyteContext
+from flytekit.core.task_executor import FlyteTaskExecutor
 from flytekit.engines.flyte import engine as _flyte_engine
 from flytekit.models import common as _common_model
+from flytekit.models import dynamic_job as dynamic_job
+from flytekit.models import literals as literal_models
 from flytekit.models import task as _task_model
 from flytekit.models.admin import common as _admin_common
 from flytekit.models.core import identifier as _identifier_model
 
 
-class FlyteTask(_hash_mixin.HashOnReferenceMixin, _task_model.TaskTemplate):
-    def __init__(self, id, type, metadata, interface, custom, container=None, task_type_version=0, config=None):
+class FlyteTask(ExecutableTaskMixin, _hash_mixin.HashOnReferenceMixin, _task_model.TaskTemplate):
+    def __init__(
+        self,
+        id,
+        type,
+        metadata,
+        interface,
+        custom,
+        container=None,
+        task_type_version=0,
+        config=None,
+        executor: Optional[FlyteTaskExecutor] = None,
+    ):
         super(FlyteTask, self).__init__(
             id,
             type,
@@ -21,6 +40,12 @@ class FlyteTask(_hash_mixin.HashOnReferenceMixin, _task_model.TaskTemplate):
             task_type_version=task_type_version,
             config=config,
         )
+
+        self._executor = executor
+
+    @property
+    def executor(self) -> Optional[FlyteTaskExecutor]:
+        return self._executor
 
     @property
     def interface(self) -> _interfaces.TypedInterface:
@@ -93,3 +118,23 @@ class FlyteTask(_hash_mixin.HashOnReferenceMixin, _task_model.TaskTemplate):
         flyte_task = cls.promote_from_model(admin_task.closure.compiled_task.template)
         flyte_task._id = admin_task.id
         return flyte_task
+
+    def execute(self, **kwargs) -> Any:
+        """
+        This function directs execute to the executor instead of attempting to run itself.
+        """
+        if self.executor is None:
+            raise ValueError(f"Cannot execute without an executor")
+
+        return self.executor.execute_from_model(self, **kwargs)
+
+    def dispatch_execute(
+        self, ctx: FlyteContext, input_literal_map: literal_models.LiteralMap
+    ) -> Union[literal_models.LiteralMap, dynamic_job.DynamicJobSpec]:
+        """
+        This function directs execute to the executor instead of attempting to run itself.
+        """
+        if self.executor is None:
+            raise ValueError(f"Cannot run dispatch_execute without an executor")
+
+        return self.executor.dispatch_execute(ctx, self, input_literal_map)
