+++ flytekit/core/python_third_party_task.py
@@ -2,18 +2,17 @@ from __future__ import annotations
 
 import importlib
 import os
-from typing import Any, Dict, Generic, List, Optional, TypeVar, Union
+from typing import Any, Dict, List, Optional, TypeVar, Union
 
 from flyteidl.core import tasks_pb2 as _tasks_pb2
 
 from flytekit.common import utils as common_utils
 from flytekit.common.tasks.raw_container import _get_container_definition
-from flytekit.common.tasks.sdk_runnable import ExecutionParameters
-from flytekit.core.base_task import PythonTask, Task, TaskResolverMixin
+from flytekit.core.base_task import ExecutableTaskMixin, PythonTask, Task, TaskResolverMixin
 from flytekit.core.context_manager import FlyteContext, Image, ImageConfig, SerializationSettings
 from flytekit.core.resources import Resources, ResourceSpec
+from flytekit.core.task_executor import FlyteTaskExecutor
 from flytekit.core.tracker import TrackedInstance
-from flytekit.core.type_engine import TypeEngine
 from flytekit.loggers import logger
 from flytekit.models import dynamic_job as _dynamic_job
 from flytekit.models import literals as _literal_models
@@ -23,137 +22,6 @@ from flytekit.models.security import Secret, SecurityContext
 
 T = TypeVar("T")
 TC = TypeVar("TC")
-
-
-class TaskTemplateExecutor(TrackedInstance, Generic[T]):
-    @classmethod
-    def execute_from_model(cls, tt: _task_model.TaskTemplate, **kwargs) -> Any:
-        raise NotImplementedError
-
-    @classmethod
-    def pre_execute(cls, user_params: ExecutionParameters) -> ExecutionParameters:
-        """
-        This function is a stub, just here to keep dispatch_execute compatibility between this class and PythonTask.
-        """
-        return user_params
-
-    @classmethod
-    def post_execute(cls, user_params: ExecutionParameters, rval: Any) -> Any:
-        """
-        This function is a stub, just here to keep dispatch_execute compatibility between this class and PythonTask.
-        """
-        return rval
-
-    @classmethod
-    def dispatch_execute(
-        cls, ctx: FlyteContext, tt: _task_model.TaskTemplate, input_literal_map: _literal_models.LiteralMap
-    ) -> Union[_literal_models.LiteralMap, _dynamic_job.DynamicJobSpec]:
-        """
-        This function is copied from PythonTask.dispatch_execute. Will need to make it a mixin and refactor in the
-        future.
-        """
-
-        # Invoked before the task is executed
-        new_user_params = cls.pre_execute(ctx.user_space_params)
-
-        # Create another execution context with the new user params, but let's keep the same working dir
-        with ctx.new_execution_context(
-            mode=ctx.execution_state.mode,
-            execution_params=new_user_params,
-            working_dir=ctx.execution_state.working_dir,
-        ) as exec_ctx:
-            # Added: Have to reverse the Python interface from the task template Flyte interface
-            #  This will be moved into the FlyteTask promote logic instead
-            guessed_python_input_types = TypeEngine.guess_python_types(tt.interface.inputs)
-            native_inputs = TypeEngine.literal_map_to_kwargs(exec_ctx, input_literal_map, guessed_python_input_types)
-
-            logger.info(f"Invoking FlyteTask executor {tt.id.name} with inputs: {native_inputs}")
-            try:
-                native_outputs = cls.execute_from_model(tt, **native_inputs)
-            except Exception as e:
-                logger.exception(f"Exception when executing {e}")
-                raise e
-
-            logger.info(f"Task executed successfully in user level, outputs: {native_outputs}")
-            # Lets run the post_execute method. This may result in a IgnoreOutputs Exception, which is
-            # bubbled up to be handled at the callee layer.
-            native_outputs = cls.post_execute(new_user_params, native_outputs)
-
-            # Short circuit the translation to literal map because what's returned may be a dj spec (or an
-            # already-constructed LiteralMap if the dynamic task was a no-op), not python native values
-            if isinstance(native_outputs, _literal_models.LiteralMap) or isinstance(
-                native_outputs, _dynamic_job.DynamicJobSpec
-            ):
-                return native_outputs
-
-            expected_output_names = list(tt.interface.outputs.keys())
-            if len(expected_output_names) == 1:
-                # Here we have to handle the fact that the task could've been declared with a typing.NamedTuple of
-                # length one. That convention is used for naming outputs - and single-length-NamedTuples are
-                # particularly troublesome but elegant handling of them is not a high priority
-                # Again, we're using the output_tuple_name as a proxy.
-                # Deleted some stuff
-                native_outputs_as_map = {expected_output_names[0]: native_outputs}
-            elif len(expected_output_names) == 0:
-                native_outputs_as_map = {}
-            else:
-                native_outputs_as_map = {
-                    expected_output_names[i]: native_outputs[i] for i, _ in enumerate(native_outputs)
-                }
-
-            # We manually construct a LiteralMap here because task inputs and outputs actually violate the assumption
-            # built into the IDL that all the values of a literal map are of the same type.
-            literals = {}
-            for k, v in native_outputs_as_map.items():
-                literal_type = tt.interface.outputs[k].type
-                py_type = type(v)
-
-                if isinstance(v, tuple):
-                    raise AssertionError(f"Output({k}) in task{tt.id.name} received a tuple {v}, instead of {py_type}")
-                try:
-                    literals[k] = TypeEngine.to_literal(exec_ctx, v, py_type, literal_type)
-                except Exception as e:
-                    raise AssertionError(f"failed to convert return value for var {k}") from e
-
-            outputs_literal_map = _literal_models.LiteralMap(literals=literals)
-            # After the execute has been successfully completed
-            return outputs_literal_map
-
-
-class ExecutorTask(object):
-    def __init__(self, tt: _task_model.TaskTemplate, executor: TaskTemplateExecutor):
-        self._executor = executor
-        self._task_template = tt
-
-    @property
-    def task_template(self) -> _task_model.TaskTemplate:
-        return self._task_template
-
-    @property
-    def executor(self) -> TaskTemplateExecutor:
-        return self._executor
-
-    def execute(self, **kwargs) -> Any:
-        """
-        This function overrides the default task execute behavior.
-
-        Execution for third-party tasks is different from tasks that run the user workflow container.
-        1. Serialize the task out to a TaskTemplate.
-        2. Pass the template over to the Executor to run, along with the input arguments.
-        3. Executor will reconstruct the Python task class object, before running the e
-
-        When overridden for unit testing using the patch operator, all these steps will be skipped and the mocked code,
-        which should just take in and return Python native values, will be run.
-        """
-        return self.executor.execute_from_model(self.task_template, **kwargs)
-
-    def dispatch_execute(
-        self, ctx: FlyteContext, input_literal_map: _literal_models.LiteralMap
-    ) -> Union[_literal_models.LiteralMap, _dynamic_job.DynamicJobSpec]:
-        """
-        This function overrides the default task execute behavior.
-        """
-        return self.executor.dispatch_execute(ctx, self.task_template, input_literal_map)
 
 
 class PythonThirdPartyContainerTask(PythonTask[TC]):
@@ -170,8 +38,8 @@ class PythonThirdPartyContainerTask(PythonTask[TC]):
         name: str,
         task_config: TC,
         container_image: str,
-        executor: TaskTemplateExecutor,
-        task_resolver: Optional[TaskTemplateResolver] = None,
+        executor: FlyteTaskExecutor,
+        task_resolver: Optional[FlyteTaskResolver] = None,
         task_type="python-task",
         requests: Optional[Resources] = None,
         limits: Optional[Resources] = None,
@@ -224,7 +92,7 @@ class PythonThirdPartyContainerTask(PythonTask[TC]):
         # Because instances of these tasks rely on the task template in order to run even locally, we'll cache it
         self._task_template = None
 
-        self._task_resolver = task_resolver or default_task_template_resolver
+        self._task_resolver = task_resolver or default_flyte_task_resolver
 
     def get_custom(self, settings: SerializationSettings) -> Dict[str, Any]:
         # Overriding base implementation to raise an error, force third-party task author to implement
@@ -243,11 +111,11 @@ class PythonThirdPartyContainerTask(PythonTask[TC]):
         return self._resources
 
     @property
-    def executor(self) -> TaskTemplateExecutor:
+    def executor(self) -> FlyteTaskExecutor:
         return self._executor
 
     @property
-    def task_resolver(self) -> TaskTemplateResolver:
+    def task_resolver(self) -> FlyteTaskResolver:
         return self._task_resolver
 
     @property
@@ -349,14 +217,16 @@ def load_executor(object_location: str) -> Any:
     return getattr(class_obj_mod, class_obj_key)
 
 
-class TaskTemplateResolver(TrackedInstance, TaskResolverMixin):
+class FlyteTaskResolver(TrackedInstance, TaskResolverMixin):
     def __init__(self):
-        super(TaskTemplateResolver, self).__init__()
+        super(FlyteTaskResolver, self).__init__()
 
     def name(self) -> str:
         return "task template resolver"
 
-    def load_task(self, loader_args: List[str]) -> Task:
+    def load_task(self, loader_args: List[str]) -> ExecutableTaskMixin:
+        from flytekit.control_plane.tasks.task import FlyteTask
+
         logger.info(f"Task template loader args: {loader_args}")
         ctx = FlyteContext.current_context()
         task_template_local_path = os.path.join(ctx.execution_state.working_dir, "task_template.pb")
@@ -365,7 +235,9 @@ class TaskTemplateResolver(TrackedInstance, TaskResolverMixin):
         task_template_model = _task_model.TaskTemplate.from_flyte_idl(task_template_proto)
 
         executor = load_executor(loader_args[1])
-        return ExecutorTask(task_template_model, executor)
+        ft = FlyteTask.promote_from_model(task_template_model)
+        ft._executor = executor
+        return ft
 
     def loader_args(self, settings: SerializationSettings, t: PythonThirdPartyContainerTask) -> List[str]:
         return ["{{.taskTemplatePath}}", f"{t.executor.__module__}.{t.executor.__name__}"]
@@ -374,4 +246,4 @@ class TaskTemplateResolver(TrackedInstance, TaskResolverMixin):
         return []
 
 
-default_task_template_resolver = TaskTemplateResolver()
+default_flyte_task_resolver = FlyteTaskResolver()
