+++ flytekit/extras/external_container_sqlite3/Dockerfile
@@ -4,4 +4,4 @@ ENV FLYTE_INTERNAL_IMAGE=flytecli-sqlite3:v0.1.0
 
 RUN pip install awscli
 
-RUN pip install -U https://github.com/flyteorg/flytekit/archive/4089babbdc7797d8e12203bda4c83c172f1c5758.zip#egg=flytekit
+RUN pip install -U https://github.com/flyteorg/flytekit/archive/ceb80db9a09928f64afded5fe9ff46ce7cce03fd.zip#egg=flytekit
