+++ tests/flytekit/unit/extras/external_container_sqlite3/test_task.py
@@ -88,10 +88,10 @@ def test_task_serialization():
         "--raw-output-data-prefix",
         "{{.rawOutputDataPrefix}}",
         "--resolver",
-        "flytekit.core.python_third_party_task.default_task_template_resolver",
+        "flytekit.core.python_third_party_task.default_flyte_task_resolver",
         "--",
         "{{.taskTemplatePath}}",
-        "flytekit.extras.sqlite3.task.SQLite3TaskExecutor",
+        "flytekit.extras.external_container_sqlite3.task.SQLite3TaskExecutor",
     ]
 
     assert tt.custom["query_template"] == "select TrackId, Name from tracks limit {{.inputs.limit}}"
